Add CaliperSession interface, AbstractSession, Session extending AbstractSession, LtiSession extending Session.  Resolves #343.
