+++ src/main/java/org/imsglobal/caliper/entities/session/CaliperSession.java
@@ -1,0 +1,57 @@
+/**
+ * This file is part of IMS Caliper Analytics™ and is licensed to
+ * IMS Global Learning Consortium, Inc. (http://www.imsglobal.org)
+ * under one or more contributor license agreements.  See the NOTICE
+ * file distributed with this work for additional information.
+ *
+ * IMS Caliper is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation, version 3 of the License.
+ *
+ * IMS Caliper is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+ * A PARTICULAR PURPOSE.  See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package org.imsglobal.caliper.entities.session;
+
+import org.imsglobal.caliper.entities.CaliperEntity;
+import org.imsglobal.caliper.entities.agent.CaliperAgent;
+import org.joda.time.DateTime;
+
+/**
+ * The session interface marks an object type that represents a user session.  The interface allows
+ * Caliper to catch errors at compile time for instances of the marked class rather than
+ * at runtime.
+ */
+public interface CaliperSession extends CaliperEntity {
+
+    /**
+     * Agent, typically a Person, who initiated the Session.
+     * @return Agent associated with this Session.
+     */
+    CaliperAgent getUser();
+
+    /**
+     * A combined date and time representation (including milliseconds) indicating the Session start time, formatted
+     * in accordance with the ISO 8601 standard.  Optional.
+     * @return
+     */
+    DateTime getStartedAtTime();
+
+    /**
+     * A combined date and time representation (including milliseconds) indicating the Session end time, formatted
+     * in accordance with the ISO 8601 standard.  Optional.
+     * @return
+     */
+    DateTime getEndedAtTime();
+
+    /**
+     * Session duration formatted in accordance with the ISO 8601 standard.  Optional.
+     * @return
+     */
+    String getDuration();
+}
