+++ src/main/java/org/imsglobal/caliper/entities/session/Session.java
@@ -29,110 +29,24 @@ import org.joda.time.DateTime;
 
 import javax.annotation.Nullable;
 
-public class Session extends AbstractEntity {
-
-    @JsonProperty("user")
-    private final CaliperAgent user;
-
-    @JsonIgnore
-    private TimePeriod timePeriod = new TimePeriod();
+public class Session extends AbstractSession {
 
     /**
      * @param builder apply builder object properties to the Session object.
      */
-    protected Session(Builder<?> builder) {
-        super(builder);
-
-        EntityValidator.checkStartTime(builder.timePeriod.getStartedAtTime(), builder.timePeriod.getEndedAtTime());
-        EntityValidator.checkDuration(builder.timePeriod.getDuration());
-
-        this.user = builder.user;
-        this.timePeriod.setStartedAtTime(builder.timePeriod.getStartedAtTime());
-        this.timePeriod.setEndedAtTime(builder.timePeriod.getEndedAtTime());
-        this.timePeriod.setDuration(builder.timePeriod.getDuration());
-    }
-
-    /**
-     * @return the session user
-     */
-    @Nullable
-    public CaliperAgent getUser() {
-        return user;
-    }
-
-    /**
-     * @return started at time
-     */
-    @Nullable
-    public DateTime getStartedAtTime() {
-        return timePeriod.getStartedAtTime();
-    }
-
-    /**
-     * @return ended at time
-     */
-    @Nullable
-    public DateTime getEndedAtTime() {
-        return timePeriod.getEndedAtTime();
-    }
-
-    /**
-     * @return duration
-     */
-    @Nullable
-    public String getDuration() {
-        return timePeriod.getDuration();
-    }
+    protected Session(Builder<?> builder) { super(builder); }
 
     /**
      * Builder class provides a fluid interface for setting object properties.
      * @param <T> builder
      */
-    public static abstract class Builder<T extends Builder<T>> extends AbstractEntity.Builder<T>  {
-        private CaliperAgent user;
-        private TimePeriod timePeriod = new TimePeriod();
+    public static abstract class Builder<T extends Builder<T>> extends AbstractSession.Builder<T>  {
 
         /**
          * Initialize type with default value.
          */
         public Builder() {
             super.type(EntityType.SESSION);
-        }
-
-        /**
-         * @param user
-         * @return builder.
-         */
-        public T user(CaliperAgent user) {
-            this.user = user;
-            return self();
-        }
-
-        /**
-         * @param startedAtTime
-         * @return
-         */
-        public T startedAtTime(DateTime startedAtTime) {
-            this.timePeriod.setStartedAtTime(startedAtTime);
-            return self();
-        }
-
-        /**
-         * @param endedAtTime
-         * @return builder
-         */
-        public T endedAtTime(DateTime endedAtTime) {
-            this.timePeriod.setEndedAtTime(endedAtTime);
-            return self();
-        }
-
-        /**
-         * @param duration
-         * @return
-         */
-        public T duration(String duration) {
-            this.timePeriod.setDuration(duration);
-            return self();
         }
 
         /**
